label_test_py/requirements.txt
json
pillow